fix initialization errors
