+++ acceptance-tests/test/end-to-end/tests/CCPB_SearchCaseWithCCD_test.js
@@ -1,11 +1,6 @@
 const CCPBATConstants = require('./CCPBAcceptanceTestConstants');
 
 const testConfig = require('./config/CCPBConfig');
-const bulkScanApiCalls = require("../helpers/utils");
-const stringUtils = require("../helpers/string_utils");
-const CCPBConstants = require("./CCPBAcceptanceTestConstants");
-const miscUtils = require("../helpers/misc");
-const searchCase = require("../pages/case_search");
 
 Feature('CC Pay Bubble Acceptance Tests').retry(CCPBATConstants.defaultNumberOfRetries);
 
@@ -37,42 +32,10 @@ Scenario('Amount Due case for Telephony flow', async({ I }) => {
   I.Logout();
 }).tag('@nightly @pipeline @crossbrowser');
 
-Scenario('Partially paid (Upfront remission) case for Telephony flow', async({ I, AddFees, FeesSummary, Remission }) => {
+Scenario('Partially paid (Upfront remission) case for Telephony flow', async({ I }) => {
   I.login(testConfig.TestDivorceCaseWorkerUserName, testConfig.TestDivorceCaseWorkerPassword);
   I.wait(CCPBATConstants.fiveSecondWaitTime);
-  const ccdNumber = await bulkScanApiCalls.createACCDCaseForDivorce();
-  const ccdCaseNumberFormatted = stringUtils.getCcdCaseInFormat(ccdNumber);
-  await miscUtils.multipleSearch(searchCase, I, ccdCaseNumberFormatted);
-  I.wait(CCPBATConstants.fiveSecondWaitTime);
-  I.see('Case transactions');
-  I.see('Case reference:');
-  I.see(ccdCaseNumberFormatted);
-  // this.click('Take telephony payment');
-  I.click('Create service request and pay');
-  I.wait(CCPBConstants.fiveSecondWaitTime);
-  AddFees.addFeesAmount('593.00', 'family', 'family_court');
-  FeesSummary.verifyFeeSummaryTelephonyPayment(ccdCaseNumberFormatted, 'FEE0002', '593.00', false);
-  FeesSummary.deductRemission();
-  Remission.processRemission('FEE0002', '200');
-  Remission.confirmProcessRemission();
-  I.wait(CCPBATConstants.fiveSecondWaitTime);
-  I.click('Case Transaction');
-  I.wait(CCPBConstants.fiveSecondWaitTime);
-  await miscUtils.multipleSearch(searchCase, I, ccdNumber);
-  I.wait(CCPBATConstants.fiveSecondWaitTime);
-  I.see('Case transactions');
-  I.see('Case reference:');
-  I.see(ccdCaseNumberFormatted);
-  I.see('Partially paid');
-  I.click('Take telephony payment');
-  I.wait(CCPBConstants.fiveSecondWaitTime);
-  FeesSummary.verifyFeeSummaryAfterRemission('FEE0002', '£593.00', '£393.00', '£200.00');
-  I.click('Take payment');
-  I.wait(CCPBConstants.fiveSecondWaitTime);
-  I.waitInUrl('pcipal', 2);
-  I.click('Cancel');
-  I.click('Finish Session');
-  I.wait(CCPBConstants.fiveSecondWaitTime);
+  await I.partiallyPaidUpfrontRemissionCaseForTelephonyFlow();
   I.Logout();
 }).tag('@nightly @pipeline');
